Bugfix: Fix bug where \`web:export-world\` would only ever error
And, in the process, remove the dependency on Akka, use futures, and
greatly simplify the code.

The source of the problem was that Akka was having a tough time finding
its configuration file.  Easier than figuring out how to tell it where
to get that file is to just remove the actor cruft altogether.